+++ clients/src/components/ssNav.jsx
@@ -9,14 +9,21 @@ class SSNav extends React.Component {
       super(props)
   
       this.state = {
+        loggedIn: false
       }
     }
     
+    componentDidMount() {
+      this.setState({
+        loggedIn: this.getCookie("refresh_token")
+      })
+    }
+
     render = () => {
       return (
         <Navbar expand="xl" className="color-nav" variant="light">
         <Container className = "navcontainer">
-          <Navbar.Brand className="extend" href = "/home">
+          <Navbar.Brand className="extend" href = {this.state.loggedIn ? "/home": "/"}>
               <img
               alt=""
               src={Logo}
@@ -28,8 +35,10 @@ class SSNav extends React.Component {
           </Navbar.Brand>
           <Nav className="mr-auto">
               <Nav.Link href="/about">About</Nav.Link>
-          </Nav>
-          <Button id = "logout" variant= "dark" size= "sm" onClick={() => this.logOut()}>Logout</Button>{' '}
+          </Nav>{
+            this.state.loggedIn ?
+          <Button id = "logout" variant= "dark" size= "sm" onClick={() => this.logOut()}>Logout</Button> : 
+          <Button id = "logout" variant= "dark" size= "sm" href="https://smartshuffle.io/login">Login with Spotify</Button>}
           </Container>
       </Navbar>
       )
@@ -38,7 +47,23 @@ class SSNav extends React.Component {
     logOut = () => {
         document.cookie = "refresh_token=; expires=Thu, 01 Jan 1970 00:00:00 UTC; path=/;";
         window.location.href = '/'
+    }
+
+    getCookie(cname) {
+      var name = cname + "=";
+      var decodedCookie = decodeURIComponent(document.cookie);
+      var ca = decodedCookie.split(';');
+      for(var i = 0; i <ca.length; i++) {
+        var c = ca[i];
+        while (c.charAt(0) === ' ') {
+          c = c.substring(1);
+        }
+        if (c.indexOf(name) === 0) {
+          return c.substring(name.length, c.length);
+        }
       }
+      return "" 
+    }
 
   }